requirements.txt
# frappe -- https://github.com/frappe/frappe is installed via 'bench init'
qrcode
pillow